+++ spec/models/user_spec.rb
@@ -1,0 +1,12 @@
+require "rails_helper"
+
+RSpec.describe User, type: :model do
+  describe "validations" do
+    it { should validate_presence_of(:name) }
+    it { should validate_presence_of(:email) }
+  end
+
+  describe "associations" do
+    it { should have_and_belong_to_many(:organisations) }
+  end
+end