+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 
 - Add Google Tag Manager in place of templated Google Analytics code
 - Ensure missing I18n strings cause tests to fail
+- Users can be created both locally and in Auth0
+- Users can be associated with multiple organisations